Refactor how a token is issued
Both author and reader accounts will receive a token.
We tell the difference between author and reader accounts, by adding
an isAuthor property in the token payload.
This change is necessary because we are going to use
passport.authenticate and JwtStrategy to assign to req.user.
Therefore, both author and reader accounts would need to have their own
individual token.